tensorflow: Add support for distributed tensorflow samples
In addition to general cleanup.  Templates and playbooks have been added to
support setting up parameter and worker servers for distributed tensorflow.

The playbook will read the config file and determine how many workers and
parameter servers are defined, and build each distributed workload accordingly.

At the moment we support the cifar10 and mnist tests, with plans to move on to
inception, as well as user-defined workloads.